Create Combine code for heart rate+GPS+Oximeter.ino
This code integrates functionalities for heart rate, GPS, and oximeter. It was developed to address issues we encountered when using the accelerometer alongside the heart rate and GPS modules, as the accelerometer was preventing the heart rate values from being correctly obtained. To resolve this, we removed the accelerometer code and integrated the oximeter instead.